feat: add initial backend for users

+++ src/services/userService.js
@@ -10,11 +10,19 @@ export class UserService {
     }
 
     async createUser(user) {
+        const existingUser = await this.getUserById(user.id_user);
+        if (existingUser) {
+            throw new Error('User already exists');
+        }
         const newUser = new UserModel(user);
         return newUser.save();
     }
 
     async deleteUser(id_user) {
+        const user = await this.getUserById(id_user);
+        if (!user) {
+            throw new Error('User not found');
+        }
         return UserModel.deleteOne({ id_user: id_user });
     }
 
@@ -34,6 +42,9 @@ export class UserService {
         const user = await this.getUserById(id_user);
         if (!user) {
             throw new Error('User not found');
+        }
+        if (user.imageUrls.length <= 0) {
+            throw new Error('No images to remove');
         }
         const index = user.imageUrls.indexOf(imageUrl);
         if (index === -1) {
